% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bus_route.R
\name{get_bus_route}
\alias{get_bus_route}
\title{Get bus route information}
\usage{
get_bus_route(routeName, from = "IETT")
}
\arguments{
\item{routeName}{Route name}

\item{from}{Website the data fetched ("trafi","IETT")}
}
\value{
List object
        route.going Route shape of going
        route.return Route shape of return
        stops.going Bus stops of going
        stops.return Bus stops of return
}
\description{
Get bus route information
}
\examples{
\dontrun{
  route <- get_bus_route("43R")
  head(route$stops.going)
  plot_bus_route(route)
}
}
