% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bus_route.R
\name{plot_bus_route}
\alias{plot_bus_route}
\title{Plot bus route and stops}
\usage{
plot_bus_route(route)
}
\arguments{
\item{route}{}
}
\description{
Plot bus route and stops
}
\examples{
\dontrun{
  route <- get_bus_route("43R")
  plot_bus_route(route)
}
}
